Fix pinFailTime not saving to keychain

+++ breadwallet/src/WalletManager+Auth.swift
@@ -176,7 +176,7 @@ extension WalletManager : WalletAuthenticator {
                 }
                 
                 if try secureTime > keychainItem(key: keychainKey.pinFailTime) ?? 0 {
-                    try setKeychainItem(key: keychainKey.pinFailTime, item: secureTime)
+                    try setKeychainItem(key: keychainKey.pinFailTime, item: Int64(secureTime))
                 }
             }
